+++ app/scripts/directives/timeline-directives.js
@@ -23,20 +23,21 @@ angular.module('lizard-nxt')
 
     var dimensions = {
       width: window.innerWidth,
-      height: 35,
+      height: 75,
       events: 20,
-      bars: 25,
+      bars: 35,
       padding: {
-        top: 5,
+        top: 12,
         right: 30,
-        bottom: 15,
+        bottom: 20,
         left: 30
       }
     },
+
     start = scope.timeState.start,
     end = scope.timeState.end,
 
-    el = element[0].getElementsByTagName('svg')[0],
+    el = element.find('svg'),
 
     interaction = {
 
@@ -99,12 +100,9 @@ angular.module('lizard-nxt')
     // keep track of events in this scope
     scope.events = {events: 0, slugs: []};
 
-    // Move timeline element into sight
-    d3.select(element[0]).transition().duration(300).style('bottom', 0);
-
     // Initialise timeline
     var timeline = new Timeline(
-      el, dimensions, start, end, interaction, scope.events.nEvents);
+      el[0], dimensions, start, end, interaction, scope.events.nEvents);
 
     // Activate zoom and click listener
     timeline.addZoomListener();
@@ -123,15 +121,16 @@ angular.module('lizard-nxt')
      */
     var updateTimelineHeight = function (newDim, dim, nEventTypes) {
       var eventHeight;
-      if (scope.mapState.getActiveTemporalLayerGroup()) {
-        eventHeight = nEventTypes * dim.events;
+      if (scope.mapState.getActiveTemporalLayerGroup() && nEventTypes > 0) {
+        eventHeight = (nEventTypes - 2) * dim.events;
         eventHeight = eventHeight > 0 ? eventHeight : 0; // Default to 0px
         newDim.height = dim.height + dim.bars + eventHeight;
       } else {
-        eventHeight = (nEventTypes) * dim.events;
+        eventHeight = (nEventTypes - 2) * dim.events;
         eventHeight = eventHeight > 0 ? eventHeight : 0; // Default to 0px
         newDim.height = dim.height + eventHeight;
       }
+
       timeline.resize(newDim,
                       scope.timeState.at,
                       scope.timeState.aggWindow,